update config file

+++ src/main/java/codeverse/com/web_be/service/GenericServiceImpl.java
@@ -1,5 +1,7 @@
 package codeverse.com.web_be.service;
 
+import codeverse.com.web_be.exception.AppException;
+import codeverse.com.web_be.exception.ErrorCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -35,6 +37,12 @@ public abstract class GenericServiceImpl<T, ID> implements IGenericService<T, ID
 
     @Override
     public void deleteById(ID id) {
+        if(!existsById(id)) throw new AppException(ErrorCode.NOT_EXISTED_ENTITY);
         repository.deleteById(id);
+    }
+
+    @Override
+    public boolean existsById(ID id) {
+        return repository.existsById(id);
     }
 }
